feat[#62] : 클라이언트 상태 로직 컴포넌트 내로 이동
