+++ src/features/auth/api/login.ts
@@ -1,5 +1,4 @@
 import { useQuery, useMutation } from "@tanstack/react-query";
-import { useAuthStore } from "@/shared/store/auth";
 import { LOGIN_END_POINT } from "../constants";
 
 interface LoginResponse {
@@ -54,9 +53,6 @@ interface EmailLoginFormData {
 
 // TODO 리액트 쿼리를 활용하여 최적화 하기
 export const usePostLoginForm = () => {
-  const setToken = useAuthStore((state) => state.setToken);
-  const setRole = useAuthStore((state) => state.setRole);
-
   const mutate = useMutation<LoginResponse, Error, EmailLoginFormData>({
     mutationFn: async (formData) => {
       const response = await fetch(LOGIN_END_POINT.EMAIL, {
@@ -77,17 +73,6 @@ export const usePostLoginForm = () => {
       }
 
       return data;
-    },
-    onSuccess: (data) => {
-      const { token, role } = data.content;
-      // 성공하면 토큰을 저장하기
-      setToken(token);
-      setRole(role);
-      // TODO 이전에 접속했던 페이지로 리다이렉션 시키기
-    },
-    // TODO 에러 핸들링 로직 작성하기
-    onError: (error) => {
-      alert(error.message);
     },
   });
