+++ src/features/auth/ui/LoginForm.tsx
@@ -1,17 +1,19 @@
 import { EmailInput, PasswordInput } from "@/entities/auth/ui";
 import { Button } from "@/shared/ui/button";
+import { useAuthStore } from "@/shared/store/auth";
 import { useLoginForm } from "../model";
 import { usePostLoginForm } from "../api";
 
 const LoginForm = () => {
   const { loginForm, handler, formValidationResult } = useLoginForm();
+  const setToken = useAuthStore((state) => state.setToken);
+  const setRole = useAuthStore((state) => state.setRole);
+  const { mutate: postLoginForm } = usePostLoginForm();
 
   const { email, password, persistLogin } = loginForm;
   const { emailHasError, emailIsEmpty, emailStatusText } = formValidationResult;
   const { handleChange, handlePersistLogin } = handler;
   const passwordIsEmpty = password.length === 0;
-
-  const { mutate: postLoginForm } = usePostLoginForm();
 
   return (
     <form
@@ -23,7 +25,21 @@ const LoginForm = () => {
           alert("아이디 또는 비밀번호를 모두 입력해 주세요");
           return;
         }
-        postLoginForm({ email, password, persistLogin });
+        postLoginForm(
+          { email, password, persistLogin },
+          {
+            onSuccess: (data) => {
+              const { token, role } = data.content;
+              setToken(token);
+              setRole(role);
+              // TODO 리다이렉션하기
+            },
+            onError: (error) => {
+              // TODO : alert 창 모달로 변경하기
+              alert(error.message);
+            },
+          },
+        );
       }}
     >
       <EmailInput
